# openapi/cloudcommander-v1.yaml
openapi: 3.0.3
info:
  title: CloudCommander API
  version: 1.1.0
  description: Command/query API for CloudCommander deterministic event core.
servers:
  - url: https://api.cloudcommander.example.com
security:
  - bearerAuth: []

paths:
  /api/v1/commands/resource-allocation:
    post:
      summary: Request a node resource allocation change
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceAllocationCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/commands/dependency-edge:
    post:
      summary: Propose a dependency edge
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DependencyEdgeCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/commands/rollback:
    post:
      summary: Initiate rollback for a target aggregate to a target sequence
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - $ref: '#/components/parameters/XActorId'
        - $ref: '#/components/parameters/XActorClaims'
        - $ref: '#/components/parameters/XIdempotencyKey'
        - $ref: '#/components/parameters/XExpectedVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackCommand'
      responses:
        '202':
          description: Accepted and appended to event log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '403':
          description: Actor claim not authorized for target action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '409':
          description: Concurrency or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '422':
          description: Validation or state-transition failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '423':
          description: Aggregate is frozen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/projections/service-graph:
    get:
      summary: Return current projected service graph for tenant
      parameters:
        - $ref: '#/components/parameters/XTenantId'
      responses:
        '200':
          description: Projected service graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceGraphProjection'

  /api/v1/projections/nodes/{node_id}:
    get:
      summary: Return projected node state
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: path
          name: node_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Node projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNodeProjection'
        '404':
          description: Projection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'

  /api/v1/telemetry/system/backpressure:
    get:
      summary: Return write-path backpressure telemetry
      responses:
        '200':
          description: Backpressure state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackpressureTelemetry'

  /api/v1/telemetry/nodes/{node_id}/guardrail-state:
    get:
      summary: Return latest guardrail alert for node
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: path
          name: node_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Guardrail telemetry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailState'

  /api/v1/telemetry/events/recent:
    get:
      summary: Return latest control-loop events
      parameters:
        - $ref: '#/components/parameters/XTenantId'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Recent events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecentControllerEvent'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    XTenantId:
      in: header
      name: X-Tenant-Id
      required: true
      schema:
        type: string
        format: uuid
    XActorId:
      in: header
      name: X-Actor-Id
      required: true
      schema:
        type: string
        minLength: 1
    XActorClaims:
      in: header
      name: X-Actor-Claims
      required: false
      schema:
        type: string
    XIdempotencyKey:
      in: header
      name: X-Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
    XExpectedVersion:
      in: header
      name: X-Expected-Version
      required: true
      schema:
        type: integer
        minimum: 0

  schemas:
    ResourceAllocationCommand:
      type: object
      additionalProperties: false
      required:
        - node_id
        - target_cpu_cores
        - target_memory_gb
        - reason_code
      properties:
        node_id:
          type: string
          format: uuid
        target_cpu_cores:
          type: number
          minimum: 0.1
          maximum: 128.0
        target_memory_gb:
          type: number
          minimum: 0.5
          maximum: 1024.0
        reason_code:
          type: string
          minLength: 1
          maxLength: 50

    DependencyEdgeCommand:
      type: object
      additionalProperties: false
      required:
        - source_node_id
        - target_node_id
      properties:
        source_node_id:
          type: string
          format: uuid
        target_node_id:
          type: string
          format: uuid
        edge_type:
          type: string
          enum: [sync, async]
          default: sync

    RollbackCommand:
      type: object
      additionalProperties: false
      required:
        - target_aggregate_id
        - target_sequence_id
        - reason_code
      properties:
        target_aggregate_id:
          type: string
          format: uuid
        target_sequence_id:
          type: integer
          minimum: 0
        reason_code:
          type: string
          minLength: 1
          maxLength: 100

    EventResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - event_id
      properties:
        status:
          type: string
          enum: [accepted]
        event_id:
          type: string
          format: uuid

    ErrorMap:
      type: object
      additionalProperties: false
      required:
        - detail
      properties:
        detail:
          type: string

    ServiceGraphProjection:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - version
        - nodes
        - edges
      properties:
        tenant_id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 0
        nodes:
          type: array
          items:
            type: string
            format: uuid
        edges:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - source
              - target
            properties:
              source:
                type: string
                format: uuid
              target:
                type: string
                format: uuid

    ResourceNodeProjection:
      type: object
      additionalProperties: false
      required:
        - node_id
        - lifecycle_state
        - cpu_cores
        - memory_gb
        - last_sequence_id
        - schema_version
      properties:
        node_id:
          type: string
          format: uuid
        lifecycle_state:
          type: string
          enum: [active, orphaned, tombstoned, frozen]
        cpu_cores:
          type: number
          minimum: 0
          maximum: 128.0
        memory_gb:
          type: number
          minimum: 0
          maximum: 1024.0
        last_sequence_id:
          type: integer
          minimum: 0
        schema_version:
          type: integer
          minimum: 1

    BackpressureTelemetry:
      type: object
      additionalProperties: false
      required:
        - status
        - utilization_rho
        - arrival_rate_hz
        - service_rate_hz
        - limit_rho
      properties:
        status:
          type: string
          enum: [healthy, overloaded]
        utilization_rho:
          type: number
        arrival_rate_hz:
          type: number
        service_rate_hz:
          type: number
        limit_rho:
          type: number

    GuardrailState:
      type: object
      additionalProperties: false
      required:
        - severity
        - metric_value
        - reason
      properties:
        severity:
          type: string
        metric_value:
          type: number
        reason:
          type: string
        timestamp_utc_ms:
          type: integer
          minimum: 0

    RecentControllerEvent:
      type: object
      additionalProperties: false
      required:
        - event_id
        - event_type
        - payload
        - timestamp_utc_ms
      properties:
        event_id:
          type: string
          format: uuid
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        timestamp_utc_ms:
          type: integer
          minimum: 0